feat: add pluggable error categorization and improve docs
